Fix parsing of process start time.
So getting the start time from the /proc/pid's cration time turned out too naive. But new attempts
to use the procinfo crate turned out frustrating as well due to bugs. So here's a hand-rolled
implementation. It may have to change when the kernel does, but it's unit-tested against ps and
should be fast.