Fix license text should have been GPLv3, not LGPLv3.
The original intent (as writen in README.md) was for GPL rather than LGPL, but I copied the wrong
license text by mistake. There doesn't seem to be much point for a standalone utility to be LGPL
anyway. I guess now is the right time to "relicense" (depending on wether README.md or LICENSE.txt
is considered binding) the code, before Emlop gets hundreds of contributors and millions of users :p