parser: refactor PretendParser and HistParser into one unified struct.
Code is smaller and more readable, and should be easyer to extend and maintain.

One downside is that we now have to include a catch-all when matching event types, which might hide
real bugs and adds code noise. Another is that passing the arguments to Parser::new_* is very wordy,
but it should get cleaner with StructOpt and it's nice that Parser doesn't know anything about File
or Stdin anymore.