parser: 2x speedup by switching from regex to manual.
New code feels a bit like C, but is still very readable and the speedup is well worth it :)

It's in theory more fragile than regex parsing, so I added a few unittests. They test hypothetical
data malformations that should never happen in the wild, but the original parser ignored these
malformations as expected while the initial implementation of the new one panicked on them. The new
tests ensure that we have no regression.