Properly define exit codes, and log errors instead of panicking.
Exit codes are now defined as 0 (all fine), 1 (error), or 2 (no data found). I would have swaped
code 1 and 2, but that required changing clap's default and was a lot of work for little
gain. Errors are things like bad cli flags, or log file not found. Panicking is still the right
thing to do in case of bugs (as opposed to user error).

Started timid use of the failure crate, and properly propagating Result::Err.