Add colored output.
Just eye candy but it does make emlop nicer to use, and we wouldn't want to compare badly against
other tools :p

The implementation was a bit less straight-forward than I hoped (many implementaions to choose from,
issues with alignment, no disable functionality in the chosen implementation) but isn't too bad. I
went with ansi_term because we have it as a sub-dep of clap and it looked friendlyer than termcolor
(sub-dep of stderrlog, which is more replaceable than clap).