predict: print estimate and elapsed even for unknown ebuilds.
Fixes #8. No new unittest because it'd require running emerge, but the existing unittest is updated.

The estimate is by definition unknown, so we print '?'. Printing nothing seem too easy to misread
(especially if that's the only merge in the output), and printing '-1' is more obviously wrong than
with the list subcommand.