parser: run in a separate thread.
This speeds up most operations. We're now consistently faster than qlop for large outputs. There's
probably a way to parse using all cores and be even faster, but using just one extra thread is easy
and already very beneficial.

The code structure has been revamped a bit, going back to a less object-oriented design. Still very
readable (better ?), actually a bit smaller, but most importantly was easyer to reason with. The
threading part was easy but passing the filter closures turned out to be surprisingly difficult to
get right. The chosen solution uses `impl Trait`, which raises our requirement to Rust-1.26. Not
stable in portage yet, but we're in no rush to release.