stats: Optionaly show totals.
Like Sync, the columns are total time, count, and average time. Will need a `--headers` option to
make that clearer.

There's a CLI papercut: the `total` option appears to be valid for both `stats` and `list` but it's
only used for `stats`. Fit fix later.

Fixed unittests (new cli option and amount of whitespace), and added some tests for --show
combinations.