feat: Show predicted merge datetime.
We were already showing the predicted merge duration, but the datetime is arguably more
useful. It'll also make it easyer to check if the early prediction was good or not.

One drawback of this feature is that unit-testing it is racy. There's a hack in place to reduce the
failure window, but it'll probably only work on the second run of the test, when compilation has
been cached.