feat: Initial support for displaying unmerge events.
Just the basics for now: no doc/unittest, 'log' mode only, and the output format is identical to
merges, but this will have to change.

The amount of space varies a bit in emerge.log for unmerge events, so this includes a small refactor
to skip whitespace for all event types.