fix: Avoid panic when encountering short emerge process commandline
Affected ~1 in 100 `emlop p` runs, so was hard to get a handle on. Once diagnosed it was easy to
artificially reproduce and fix, but I still wonder what real-world commandline tiggers it.

This is a quick-fix for the bug, but it would be nice to avoid the `...` when unnecessary and to
take the formated pid lengh into account.