output: Format dates with `time` crate and respect `--utc`
This change causes a regression in the handling of DST timezones : the current utc offset will be
used to display all dates, ignoring DST. The proper fix requires changes in the `time` crate, see
comments on the `timezone` test. Hopefully this incorrect DST handling is acceptable, compared to
keeping the vulnerable `chrono` crate.

Dates formatted for logs and unittests are always UTC and will not have a configurable formatting.

Some code was moved from `main.rs` to `date.rs`.