output: Replace `Tabwriter` with our own implementation
The original crate doesn't support per-column alignment, and the feature request was rejected. We
were working around this by manully aligning, but that required overfitting. We now have some fields
with a big range of lengths, so the overfitting is getting really akward. A survey of other possible
crates yielded many that seemed too featureful for our usecase. So... Let's build our own.

The impl is simple enough. API could be a bit nicer but works well. Perfromance is better than
Tabwriter (probably because code is more ad-hoc), although we're now using it for `log` as well
(which used non-tabulated writes) so there's a sligh perf regressin there. There are a handful of
non-perfect spacings, but the overall layout is improved.

This also opens up some features like headers and sorting.