deps: Drop dep on `ansi_term`
The crate is unmaintained.

I tried to use the `termcolor` crate which is already in our deps, but the API is ridiculoously
complicated for our usecase. There are other crates available, but in the end we just want 6
constants (I doubt we'll ever want to support user-configurable colors or non-ansi terms), so let's
just hardcode them and save on compilation and startup time.

Fixes: #25