predict: Better handling of colored logs
The old system could fail when an ansi escape was truncated in the middle. It also did not really
obey the --color=n setting. The new system unconditionally strips out ansi escapes. All crates I
found that implemented this had overkill features and dependencies, so here's a simple in-house
strip() that only cares about what you'd expect to find in a log file.