bench: Add a benchmark and unittest for FilterStr
Helped improving the filtering, as benching via exec_compare.rs is too noisy, and the whole-binary
unittests missed some cases.

Benchmarks are still an unstable Rust feature (and Criterion is only for libraries), so I had to
mess with features a bit to not raise the MSRV for end users. To run the benchmarks, use
`cargo +nightly bench -F unstable bench`.