config: Initial support for config file
This is still an experiment, the final feature might look very different. Loading/parsing a config
file is trivial; the hard part is merging the different sources (cli > config > default), giving
clear error feedback (should be as good as cli error feedback), and keeping boiler-plate and
performance in check. I've gone thru a few iterations to reach this MVP, there are a few
complexities left down the road, but this seems like a good snapshot to have in git histroy.