conf: Move `--from/--to` to `Conf`, use our own error type instead of clap's
Struggled for a while to get the desired output with clap errors, so I introduced our own error
type. We'll need to maintain it to match the style of clap errors, but it's not a lot of code, call
sites are cleaner, and the error strings are better (see `--from` and `--show`).