conf: Move `--resume` to `Conf`
I initially made this toml-configurable, but I don't think anybody would want to set a different
default. Added `current` enum variant to replace the use of `Option<_>`. Was needed to override a
toml setting, but still seems valuable without it.

The formatting difference between clap::Error and ArgError is not big, but I'll still want to
harmonize things somehow.