predict: Find the currently-emerging pkg using emerge's sandbox process commandline
This should be a bit more reliable than looking at the emerge logs since the oldest emerge
process. See bugs 5 and 45 for failure examples. Will need to play with this code for a while to
make sure it behaves as we want. Genlop and qlop use a similar technique, but they don't handle as
many cases and don't seem to have a fallback.