conf: Make `--color` configurable
Because I felt the need to discreetely force-enable colors for the sreencast. As a result, I need to
bring back an explicit "auto" variant. But renamed the variants to be more straightforward.

Quite happy with ArgParse here, as I can parse into a bool directly, without an enum in between and
handling the isterm logic during parsing.