bench: Add a "last sync" case
Will add it to the benchmark doc, kinda to replace teh removed `tgcc` but also because it seems like
a useful example command.  Had to add a sync as the last event of bench/emerge.log so that qlop's
`-l` option does something interesting. Genlop doesn't have a comparable flag, so using `--date`
will have to do.