completions: Improve zsh shell completion
* Add completion for ebuild names
* Ignore zsh version < 5.2 (not packaged in gentoo)

What a mess, zsh completion system is really complex, the docs is a wall of text, the tutorials
didn't help much. The clap-generated completion was buggy. This works better now, but is not as good
as bash and fish completions. I'm not a zsh user so I'll leave it at that, help welcome.