predict: Make sure we don't display a redundant emerge root proces
There's a small time window during which `emerge` might spawn `emerge`, so it looks like we have two
running. This is a longstanding emlop bug, but it's much more noticeable now that we display the
proces hierarchy. Fix this by removing roots if their (grand)parent is also a root. Added unittest
(which prompted the previous refactorings).