table: Refactor header handling
By storing the header as its own member instead of a row. This simplifies the logic a bit (might
even be a micro perf win), but more importantly will simplify inserting a `skipped` row between the
header and the actual rows. Added a unittest to make sure we handle the header's width.

As a bonus, newer clippy versions don't complain about needless_range_loop, so remove the `allow`.