log: Show skip row for `--first`
This means that `--first 1` now needs to read the entire log and is no longer as fast, but the
reature is worth it (and is optional via `--elipsis=n`).

I'm not too happy with the padding at the end of the skip row, and the inconsistency with the
`--last` skip row (no padding and no column align), but no need to be perfect at this stage.