log: Better heuristics to determine which command has terminated
This is still WIP: no unittest, probably missed corner cases, buggy unless merge are also
parsed/shown, etc. But it matches correctly the begining of my emerge.log (with lots of parallel
commands), and it seems good to commit as a snapshot.

UI-wise, `log` should have an option (maybe default) to display commands when they start rather than
when they stop, as it looks more natural. We could skip state tracking in that case.