log: Introduce `--lastmerge` option
This currently somewhat mirrors what qlop offers, as an MVP to get a feel of the implementation. I
initially wrote a single-pass version (droping previous table rows when encountering a command), but
the two-pass version is fast enough that `--lastmerge` is about as fast as `--last`, and offers more
flexibility.

Tentative todo:
* Allow `--lastmerge <N>`
* Move into `--from '1 command'` (simpler and more consistent)
* Optionaly ignore non-merge commands (qlop always/only does that, but it's a failible heuristic)
* Investigate `--to` support
* Move code into get_hist (DRY, and optimization opportunities)
* Test and document