parse/log: Initial support for binary merges
The emerge.log line that tells us whether a merge is from a binary is between the "merge start" and
"merge stop" lines, so we need to parse that new line but also need to keep a "this is a binmerge"
bool state when iterating through hist events.

This line contains other bits of info that we might want to use later, so I'm keeping new
`Hist::MergeStep` variant flexible.

This MVP just displays binary merges in a different color.