accuracy: More detailed stats
More detailed stats:
* Sign of individual errors
* Count, min, max, absolute arithmetic average, standard deviation of ratio lists
  - This replaces showing the average of per-package *errors*
  - This is now also show for the overall ratio list, giving a one-line summary

Changed `--show=...` meaning:
* per-package moved from `-t` to `p`
* summary is at `t`
* default changed from `mt` to `pt`

This was motivated by the desire for a simple overall score (issue #70), but the nature of the data
makes it hard to give a statistically sound number. I tried a few things, but in the end I settled
with this simple arithmetic mean and standard deviation of the real/estimate ratios. It must still
be taken with a grain of salt, but I think it's better than the previous numbers and satisfies the
featur request.